RendererVk: Late reordered command recording - Episode 1
This commit is known to work only on RendererDemo scene #1, up to
IndexBuffer Subscene. Further scenes are to come in a near future.

Basic PoC implementation of CommandRecorder calls reordering. Some
internal implementation details might change if this turns out to be
an invalid way to go.

+++ Src/Engine/Common/Containers/StaticArray.hpp
@@ -175,7 +175,7 @@ public:
     NFE_INLINE Iterator end() { return this->End(); }
 
 private:
-    alignas(alignof(ElementType)) uint32 mData[MaxSize * sizeof(ElementType)];
+    alignas(alignof(ElementType)) uint8 mData[MaxSize * sizeof(ElementType)];
 
     // number of elements in the array
     uint32 mSize;
